change 3: $DATADIR was not set properly and disable mpirun
TMPDIR is not set. So for now, the input data location is
hardcoded. Must be changed with the two types of datasets
mnist and cifar10 that we want to run. There should be
better solution.

mpirun was used to copy the file to all  machine local HDD.
We do not want that as we want to see the shared storage traces.
To be debugged cleanly later.

Signed-off-by: Animesh Trivedi <atrivedi@vu.nl>